make GenderPicker a ui entity
